Moved send request function to utils
